fix small problem, make middleware for isadmin, add routes, controllers, for make users posts.

+++ routes/web.php
@@ -1,32 +1,36 @@
 <?php
-
-use App\Http\Controllers\Admin\CategoryController;
-use App\Http\Controllers\Admin\CommentController;
-use App\Http\Controllers\Admin\LikeController;
-use App\Http\Controllers\Admin\PostController;
-use App\Http\Controllers\Admin\ProfileController;
-use App\Http\Controllers\Admin\RoleController;
-use App\Http\Controllers\Admin\TagController;
-use App\Http\Controllers\Admin\UserController;
-use App\Http\Controllers\MainController;
-use Illuminate\Foundation\Application;
 use Illuminate\Support\Facades\Route;
-use Inertia\Inertia;
 
 
 
-
-Route::group(['middleware' => 'auth'], function (){
-    Route::get('/', [MainController::class, 'index'])->name('dashboard');
-    Route::get('/posts', [PostController::class, 'index'])->name('posts.index');
-    Route::post('/posts', [PostController::class, 'store'])->name('posts.store');
-    Route::get('/posts/create', [PostController::class, 'create'])->name('posts.create');
-    Route::get('/posts/{post}', [PostController::class, 'show'])->name('posts.show');
-    Route::patch('/posts/{post}', [PostController::class, 'update'])->name('posts.update');
-    Route::get('/posts/{post}/edit', [PostController::class, 'edit'])->name('posts.edit');
-    Route::delete('/posts/{post}', [PostController::class, 'destroy'])->name('posts.destroy');
-    Route::get('/categories', [CategoryController::class, 'index'])->name('categories.index');
-    Route::get('/categories/create', [CategoryController::class, 'show'])->name('categories.show');
-
+Route::group(['middleware' => ['auth']], function () {
+    Route::get('/', [App\Http\Controllers\Profile\ProfileController::class, 'dashboard'])->name('dashboard');
 });
+
+// Маршруты, доступные только администраторам
+Route::group(['middleware' => ['auth', 'admin'], 'prefix' => 'admin'], function () {
+    Route::get('/posts', [App\Http\Controllers\Admin\PostController::class, 'index'])->name('admin.posts.index');
+    Route::post('/posts', [App\Http\Controllers\Admin\PostController::class, 'store'])->name('admin.posts.store');
+    Route::get('/posts/create', [App\Http\Controllers\Admin\PostController::class, 'create'])->name('admin.posts.create');
+    Route::get('/posts/{post}', [App\Http\Controllers\Admin\PostController::class, 'show'])->name('admin.posts.show');
+    Route::patch('/posts/{post}', [App\Http\Controllers\Admin\PostController::class, 'update'])->name('admin.posts.update');
+    Route::get('/posts/{post}/edit', [App\Http\Controllers\Admin\PostController::class, 'edit'])->name('admin.posts.edit');
+    Route::delete('/posts/{post}', [App\Http\Controllers\Admin\PostController::class, 'destroy'])->name('admin.posts.destroy');
+    Route::get('/categories', [App\Http\Controllers\Admin\CategoryController::class, 'index'])->name('admin.categories.index');
+    Route::get('/categories/create', [App\Http\Controllers\Admin\CategoryController::class, 'show'])->name('admin.categories.show');
+});
+
+// Маршруты для профилей, доступные всем аутентифицированным пользователям
+Route::group(['middleware' => ['auth'], 'prefix' => 'profiles'], function () {
+    Route::get('/posts', [App\Http\Controllers\Profile\PostProfileController::class, 'index'])->name('profiles.posts.index');
+    Route::post('/posts', [App\Http\Controllers\Profile\PostProfileController::class, 'store'])->name('profiles.posts.store');
+    Route::get('/posts/create', [App\Http\Controllers\Profile\PostProfileController::class, 'create'])->name('profiles.posts.create');
+    Route::get('/posts/{post}', [App\Http\Controllers\Profile\PostProfileController::class, 'show'])->name('profiles.posts.show');
+    Route::patch('/posts/{post}', [App\Http\Controllers\Profile\PostProfileController::class, 'update'])->name('profiles.posts.update');
+    Route::get('/posts/{post}/edit', [App\Http\Controllers\Profile\PostProfileController::class, 'edit'])->name('profiles.posts.edit');
+    Route::delete('/posts/{post}', [App\Http\Controllers\Profile\PostProfileController::class, 'destroy'])->name('profiles.posts.destroy');
+//    Route::get('/categories', [App\Http\Controllers\Profile\CategoryController::class, 'index'])->name('profiles.categories.index');
+//    Route::get('/categories/create', [App\Http\Controllers\Profile\CategoryController::class, 'show'])->name('profiles.categories.show');
+});
+
 require __DIR__.'/auth.php';
